# Try to speed up `#[derive(ValidGrouping)]`