While looking at https://github.com/rust-lang/rust/issues/81262, I also tried looking at the master branch to see how things had evolved since this issue was opened, in particular since the use of proc_macros.

While testing 32/64/128 columns, I noticed that:
- the macro expansion and proc_macro expansion were relatively the same duration each: indicating trying to optimize the proc_macros could be helpful.
- building `diesel` itself, it seemed most of the proc_macro slowness came from [the `TupleWrapper`s and in particular the `#[derive(ValidGrouping)]`](https://github.com/diesel-rs/diesel/blob/75b8ef3b1ed6817f12b4b9504e45cedbbb5a4fd2/diesel/src/type_impls/tuples.rs#L117-L121)

This proc_macro roughly takes one second to execute each time from 100 to 128 columns. I believe it could be interesting for compilation times to have a structure that is less expansion-heavy, esp. the `MixedAggregates` chain causing a quadratic amount of expansion.

Since I'm not familiar enough with `diesel` and its aggregates to successfully come up with such a design, I wanted to notify you about this finding, while trying to help in the meantime. This PR reduces the constant factors of this proc_macro (while still retaining its quadratic nature) so maybe it can help a bit, in that it could be "fast enough", and that new slowness and opportunities for improvement are to be found elsewhere.

It seems the mixed aggregates chain, and use of `quote` for short quoting/unquoting concatenation, builds up big token streams that are slower and slower to parse and reparse at each step. In this PR, I've tried to build up the code in a string buffer, to limit the number of times they are parsed with `quote`. It's not ideal, probably a bit more error-prone, but I don't think that `quote` was made to handle this use-case quickly (even after the [recent performance improvements](https://github.com/dtolnay/quote/releases/tag/1.0.16))

In a [benchmark with a single use of the derive](https://gist.github.com/lqd/498d6b0ae375b1fad7e47a55323d847f) per column-count, I see these results for a non-incremental `cargo check` of the benchmark, from 1 to 128 columns:

![chart - valid grouping diesel benchmark - check - 128 columns](https://user-images.githubusercontent.com/247183/166489784-b5dbfc7a-86c7-4202-b515-3f069daf9c82.png).

Proc_macro expansion itself is 30x faster at 128 columns (from 1s to 35ms or so) for this one macro invocation, and as seen in the above chart, total compilation of this simple stress test is also divided by a bit over 2 (it then also shows other bottlenecks in rustc).

That also translates to `diesel` itself, proc_macro expansion is reduced, for a `cargo check`:
- 3.5x for `32-column-tables`: from 1s to 280ms
- 10x for `64-column-tables`: from 6s to 600ms
- 22x for `128-column-tables`: from 45s to 2s

End-to-end compile times are reduced a tiny bit for the heavier cases: these -40s are somewhat noticeable, but the `cargo check` itself is around 10 minutes long, so that's only a 4-5% improvement there.

I'm opening this PR as a draft for feedback, to see if the approach is acceptable; and to see if CI likes it (as it seems likely that I forgot some use-cases with different forms of input).