@weiznich 2022-05-03
Thanks for opening this PR and doing all this analysis :heart:

While trying to put together a simple example showing why those quadratic expansion of `MixedAggregates` is required I got kind of nerd snipped in trying another approach. The historic reason of having this kind of expansion is to work around https://github.com/rust-lang/rust/issues/75542 (https://github.com/diesel-rs/diesel/issues/2480). I kind of assumed that this happens for all traits with such recursive structure, but it turns out that this is not true. 

Seems like we can replace the the corresponding derive + some other recursive impls as well. See https://github.com/weiznich/diesel/commit/61351f779a3076bfa8b8cecdea19caa560fc95fc for the corresponding change. This brings down the time of a `cargo check --features "sqlite 64-column-table" --no-default-features` build for me from ~ 1min 20s to ~1 min 05s`, so a nearly 20% improvement.

Given that success I've tried to apply the same approach to the recursive `SqlType` impl https://github.com/diesel-rs/diesel/blob/0c6219f5bd9641eea09c6fa7dca4995aefda30a0/diesel/src/type_impls/tuples.rs#L511-L560 for tuples as well. This is as far as I know one of the spots where rustc spends quite a lot of time. See https://github.com/weiznich/diesel/commit/efa182fc6895ad25d2c1a716f4452fb0040a9eaf for the corresponding change.
Unfortunately it is not possible to apply the same pattern there. This results in quite large build times (> 13 minutes for me with all `cargo check --features "sqlite" --no-default-features`). It's not clear for me why this happens and what's even different there compared to the other traits. All of those tuple impls basically work by forwarding stuff to the next smaller tuple. In therory rustc should be able to resolve that without quadratic behaviour, assuming the corresponding impls are cached.