@lqd 2022-05-04
In case it helps, here's a breakdown of the traits whose tuple impls trigger the specialization graph slowness (until I have more info on what exactly is slow there, as it's doing a bunch of things at the same time, like setting up specialization/coherence/overlap):

| trait                                                | number of impls | time in seconds |
|------------------------------------------------------|-----------------|-----------------|
| `diesel::deserialize::FromSqlRow`                    | 17              | 94              |
| `diesel::deserialize::SqlTypeOrSelectable`             | 34              | 144             |
| `diesel::deserialize::FromStaticSqlRow`           | 33              | 48              |
| `diesel::expression::TypedExpressionType`              | 36              | 47              |
| `diesel::expression::QueryMetadata`                    | 35              | 48              |
| `diesel::query_dsl::load_dsl::private::CompatibleType` | 35              | 143             |
| `diesel::sql_types::IntoNullable`                      | 18              | 126             |
| `diesel::deserialize::Queryable`                       | 49              | 95              |
| `diesel::util::TupleSize`                            | 33              | 47              |